chore: move GameObjectId to ryot_content
