+++ apps/ryot_compass/src/ui/mod.rs
@@ -2,7 +2,6 @@ use crate::sprites::LoadedSprite;
 use crate::{CursorCommand, TilesetCategory, ToolMode};
 use bevy::prelude::*;
 use egui::{Align, Ui};
-use ryot::bevy_ryot::GameObjectId;
 use ryot::prelude::*;
 use std::ops::Range;
